image

+++ src/context/imageContext/ImageProvider.js
@@ -7,37 +7,73 @@ import ImageContext from './ImageContext'
 export default class ImageProvider extends React.Component { 
 
     state = { 
-        imageList: [ 
-            {
-                name: 'name' , 
-                url: 'url', 
-                quantity: 2, 
-
-            }
-        ],
-        nameList: []
+        imageAnswerUrl: '', 
+        imageList: [ ], 
     } 
 
     async componentDidMount() { 
-          this.fetchListFromStorage()
+        //   this.fetchListFromStorage()
           this.fetchListFromFirestore()
     }
 
-    async fetchListFromStorage() {
-        const listRef = firebase.storage.ref(); 
+    // async fetchListFromStorage() {
+    //     const listRef = firebase.storage.ref(); 
         
-        let imagesArray = []
+    //     let imagesArray = []
 
-        await listRef.listAll().then(function(res) {
-            res.items.forEach(function(itemRef) { 
-                imagesArray.push(itemRef.fullPath) 
-            });
-          }).catch(function(error) {
-              console.log(error) 
-          });
-          this.setState({
-            imageList: imagesArray
-          }) 
+    //     await listRef.listAll().then(function(res) {
+    //         res.items.forEach(function(itemRef) { 
+    //             imagesArray.push(itemRef.fullPath) 
+    //         });
+    //       }).catch(function(error) {
+    //           console.log(error) 
+    //       });
+    //       this.setState({
+    //         imageList: imagesArray
+    //       }) 
+    // }
+
+    async addImageToList(newImage, oldName) {
+        const newList = this.state.imageList.slice()
+        let newItem
+        let isNew = true
+        let isNeedUpdate = newImage.name === oldName ? false : true 
+        if(isNeedUpdate) { 
+            for(let i = 0; i < newList.length; i++) {
+                if(newList[i].name === newImage.name) {
+                    //уже было загружено
+                    newList[i].quantity++ 
+                    isNew = false
+                }
+                if(newList[i].name === oldName) {
+                    //уже было загружено
+                    newList[i].quantity--  
+                    if( newList[i].quantity === 0 ) { 
+                        this.removeImage(oldName) 
+                        newList.splice(i, 1)
+                    }
+                }
+            }
+            if(isNew) {
+                newItem = {
+                    name: newImage.name, 
+                    url: newImage.url, 
+                    quantity: 1
+                }
+                newList.push(newItem)
+            }
+        }
+        else { 
+            console.log(newImage, oldName)
+        }
+         
+        this.setState({
+            imageList: newList
+        })
+ 
+        console.log(this.state.imageList, oldName)
+
+        this.updateList(newList)
     }
 
     async fetchListFromFirestore() {
@@ -49,42 +85,65 @@ export default class ImageProvider extends React.Component {
 
         } else { 
             this.setState({
-                nameList: doc.data().list
+                imageList: doc.data().list
             }) 
         }
     }
-    async uploadImage(imageData) {
-        let answerUrl
-        const storageRef = firebase.storage.ref(`${imageData.name}`).put(imageData)
-        storageRef.on('state-changed', 
-          snapshot => {
-            console.log("image uploaded")
-            // console.log( snapshot )
-          }, 
-          error => {
-            console.log(error.message)
-          },
-          () => { 
-            storageRef.snapshot.ref.getDownloadURL()
-              .then( url => {
-                answerUrl = url 
-              })
-          }
-        )
-        return answerUrl
+    async uploadImage(imageData, oldName) { 
+        const storageRef = await firebase.storage.ref(`${imageData.name}`).put(imageData)
+        const downloadURL = await storageRef.ref.getDownloadURL();
+ 
+        const newImage = {
+            url: downloadURL, 
+            name: imageData.name, 
+            quantity: 1
+        }
+        this.addImageToList(newImage, oldName)
+  
+    }
+    async updateList(newList) {
+        await firebase.db.collection("site1category").doc("imageList").update({
+            list: newList, 
+        }).then(() => { 
+            this.setState({
+                imageList: newList
+            })
+        })
     }
 
-    removeImage(name) { 
-        const storageRef = firebase.storage.ref();
+    async removeImage(name) { 
+        if(name !== '') {
+            const storageRef = firebase.storage.ref();
  
-        const imageRef = storageRef.child(name)
-     
-        imageRef.delete().then(function() {
-            console.log('File deleted successfully')
-        }).catch(function(error) {
-            console.log(error) 
-        }); 
+            const imageRef = storageRef.child(name)
+        
+            try {
+                await imageRef.delete().then(function() {
+                    console.log('File deleted successfully')
+                }).catch(function(error) {
+                    console.log(error) 
+                }); 
+            } catch (error) {
+                console.log(error)
+            }
+        }
     }
+    removeImageFromStoreAndList(name) {
+        console.log(name)
+        const newList = this.state.imageList.slice()
+        newList.map( (item, index) => {
+            if(item.name === name) {
+                item.quantity--
+                if(item.quantity === 0) newList.splice(index, 1)
+            }
+        })
+        this.setState({
+            imageList: newList
+        })  
+        this.removeImage(name)
+        this.updateList(newList)
+    }
+   
 
     render() {
         return(
@@ -92,23 +151,15 @@ export default class ImageProvider extends React.Component {
                 value={{
                     imageList: this.state.imageList, 
                     nameList: this.state.nameList ,
-                    // updateImageNameList: (name) => {
-                    //     let newList = this.state.nameList.slice()
-                    //     let isUnic = true
-                    //     newList.map((item) => {
-                    //         if( item === name ) isUnic = false
-                    //     })
-                    //     if(isUnic) newList.push(name)
-
-                    //     this.setState({
-                    //         nameList: newList
-                    //     }) 
-                    //     firebase.db.collection("site1category").doc('imageList').update({
-                    //         list: newList
-                    //     })
-                    // },
+                    
                     removeImage: (name) => {
                         this.removeImage(name)
+                    },
+                    removeImageFromStoreAndList: (name) => {
+                        this.removeImageFromStoreAndList(name)
+                    },
+                    uploadImage: (data, name) => {
+                        this.uploadImage(data, name)
                     }
                 }}
             >
